crypto/signed_identity_test.go
/* Mysterium network payment library.
 *
 * Copyright (C) 2021 BlockDev AG
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Lesser General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 * You should have received a copy of the GNU Lesser General Public License
 * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */

package crypto

import (
	"os"
	"testing"

	"github.com/stretchr/testify/assert"
)

func TestSignedIdentity(t *testing.T) {
	dir, ks := tmpKeyStore(t, false)
	defer os.RemoveAll(dir)

	account, err := ks.ImportECDSA(getPrivKey("consumer"), "")
	assert.Nil(t, err)
	if err := ks.Unlock(account, ""); err != nil {
		t.Fatal(err)
	}

	toSign := HexToAddress("0x000012312312312313")
	req, err := CreateSignedIdentityRequest(ks, toSign, FromCommonAddress(account.Address))
	assert.NoError(t, err)

	// should return ok address
	signer, err := req.RecoverSigner()
	assert.NoError(t, err)
	assert.Equal(t, account.Address, signer)
}
